small fixes
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3510 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/NewNetEngine.cpp
@@ -11,6 +11,8 @@
 #include "CClient.h"
 #include "CMap.h"
 #include "CWorm.h"
+#include "ProfileSystem.h"
+#include "CWormHuman.h"
 
 namespace NewNet
 {
@@ -66,7 +68,7 @@ bool SendNetPacket( unsigned long localTime, KeyState_t keys, CBytestream * bs )
 
 bool Frame( CBytestream * bs )
 {
-	unsigned long localTime = tLX->fCurTime * 1000;
+	unsigned long localTime = (unsigned long) (tLX->fCurTime * 1000);
 	KeyState_t keys;
 	if( cClient->getNumWorms() > 0 && cClient->getWorm(0)->getType() == PRF_HUMAN )
 	{
